+++ enlighten/network/BLEManagerFeature.py
@@ -4,6 +4,7 @@ import logging
 from queue import Queue
 
 from threading import Thread
+from datetime import datetime
 
 from enlighten import common, util
 from enlighten.EnlightenFeature import EnlightenFeature
@@ -31,6 +32,8 @@ class BLEManagerFeature(EnlightenFeature):
 
     @see detailed Bluetooth® LE architecture in wasatch.BLEDevice
     """
+
+    SECTION = "BLEManager"
 
     def __init__(self, ctl):
         super().__init__(ctl)
@@ -145,6 +148,13 @@ class BLEManagerFeature(EnlightenFeature):
         self.ble_selector.reset()
         self.ble_selector.hide()
 
+        # record this as new "last used" date
+        self.ctl.config.set(self.SECTION, device_id.serial_number, datetime.now().isoformat())
+
+        if self.ctl.resource_monitor.under_parallels():
+            self.ctl.marquee.error("BLE is not currently supported on Windows under Parallels VM")
+            return
+
         # add this DeviceID to the Controller's list of "external" (non-USB) 
         # device IDs to check
         log.debug(f"adding {device_id} to Controller external search list")
@@ -154,21 +164,22 @@ class BLESelector(QDialog):
     """
     A pop-up window listing all discovered Bluetooth® LE devices.
 
-    +-----------------------+
-    | BLE Spectrometers [x] |
-    +-----------------------+
-    | [Connect]    [Rescan] |
-    |                       |
-    | Signal Serial Number  |
-    | ***    WP-01234       |
-    | *      WP-01228       |
-    | **     WP-01499       |
-    +-----------------------+
+    +----------------------------+
+    | BLE Spectrometers      [x] |
+    +----------------------------+
+    | [Connect] [Rescan]         |
+    |                            |
+    | Signal Serial Number  Used |
+    | ***    WP-01234       1d   |
+    | *      WP-01228       2wk  |
+    | **     WP-01499            |
+    +----------------------------+
     """
     def __init__(self, ble_manager, parent=None):
         super().__init__(parent)
 
         self.ble_manager = ble_manager
+        self.ctl = self.ble_manager.ctl
 
         self.setWindowTitle("BLE Spectrometers")
         self.layout = QVBoxLayout()
@@ -179,12 +190,14 @@ class BLESelector(QDialog):
             button_layout = QHBoxLayout()
             self.bt_connect = QPushButton(text="Connect", parent=self)
             self.bt_rescan = QPushButton(text="Rescan", parent=self)
+            for b in [ self.bt_connect, self.bt_rescan ]:
+                b.setMinimumWidth(150)
             button_layout.addWidget(self.bt_connect)
             button_layout.addWidget(self.bt_rescan)
             self.layout.addLayout(button_layout)
 
-            # add QTableWidget with RSSI and Serial Number columns
-            self.table = QTableWidget(1, 2, self)
+            # add QTableWidget with RSSI, Serial Number and Last Used columns
+            self.table = QTableWidget(1, 3, self)
             self.table.horizontalHeader().setStretchLastSection(True)
             self.layout.addWidget(self.table)
 
@@ -217,6 +230,7 @@ class BLESelector(QDialog):
 
         self.table.setItem(0, 0, QTableWidgetItem("Signal"))
         self.table.setItem(0, 1, QTableWidgetItem("Serial Number"))
+        self.table.setItem(0, 2, QTableWidgetItem("Last Used"))
 
         self.serial_number_to_row = {}
         self.row_to_device_id = {}
@@ -253,6 +267,35 @@ class BLESelector(QDialog):
             self.selected_device_id = device_id
             log.debug(f"user selected {device_id}")
 
+    def human_time_unit(self, sec):
+        if not sec:
+            return
+        elif sec < 60:
+            return f"{sec} sec"
+        elif sec < 3600:
+            n = int(round(sec/60, 0))
+            return f"{n} min"
+        elif sec < 3600 * 24:
+            n = int(round(sec/(3600), 0))
+            unit = "hour" if n < 2 else "hours"
+            return f"{n} {unit}"
+        elif sec < 3600 * 24 * 7:
+            n = int(round(sec/(3600*24), 0))
+            unit = "day" if n < 2 else "days"
+            return f"{n} {unit}"
+        elif sec < 3600 * 24 * 30:
+            n = int(round(sec/(3600*24*7), 0))
+            unit = "week" if n < 2 else "weeks"
+            return f"{n} {unit}"
+        elif sec < 3600 * 24 * 365:
+            n = int(round(sec/(3600*24*30), 0))
+            unit = "month" if n < 2 else "months"
+            return f"{n} {unit}"
+        else:
+            n = int(round(sec/(3600*24*365), 0))
+            unit = "year" if n < 2 else "years"
+            return f"{n} {unit}"
+
     def add_to_table(self, discovered_device):
         """
         @param discovered_device: a wasatch.DeviceFinderBLE.DiscoveredBLEDevice
@@ -261,10 +304,20 @@ class BLESelector(QDialog):
         rssi = discovered_device.rssi
         serial_number = device_id.serial_number
 
+        # connection strength
         rssi_num_str = f"RSSI {rssi:0.2f}"
         rssi_bar_str = self.rssi_to_bars(rssi)
         rssi_item = QTableWidgetItem(rssi_bar_str)
         rssi_item.setToolTip(rssi_num_str)
+
+        # last used
+        last_used_str = self.ctl.config.get(self.ble_manager.SECTION, serial_number, default=None)
+        if last_used_str is not None:
+            last_used_date = datetime.fromisoformat(last_used_str)
+            elapsed_sec = (datetime.now() - last_used_date).total_seconds()
+            since_used = self.human_time_unit(elapsed_sec)
+        else:
+            since_used = "never"
 
         try:
             if serial_number in self.serial_number_to_row:
@@ -278,6 +331,7 @@ class BLESelector(QDialog):
 
                 self.table.setItem(row, 0, rssi_item)
                 self.table.setItem(row, 1, QTableWidgetItem(serial_number))
+                self.table.setItem(row, 2, QTableWidgetItem(since_used))
 
                 # update mappings
                 self.serial_number_to_row[serial_number] = row